add sorting to Table and refactor
