+++ libs/react/shelter-operator/src/lib/components/RoomTable.tsx
@@ -11,8 +11,8 @@ import type { CSSProperties, ReactNode } from 'react';
 import { useEffect, useMemo, useState } from 'react';
 import { Button } from './base-ui/buttons/buttons';
 import { ConfirmationModal } from './base-ui/modal/ConfirmationModal';
+import { Table, type TableColumn } from './base-ui/table';
 import { Text } from './base-ui/text/text';
-import { Table, type TableColumn } from './Table';
 
 // REPLACE WITH ACTUAL QUERIED DATA
 import {
