+++ libs/react/shelter-operator/src/lib/components/base-ui/table/index.ts
@@ -1,0 +1,9 @@
+export { Table } from './Table';
+export {
+  TableRow,
+  TableRowSlot,
+  type TableRowCell,
+  type TableRowClickHandler,
+  type TableRowProps,
+} from './TableRow';
+export type { SortDirection, TableColumn, TableProps } from './types';
